Main components of WFC should be done now.
