+++ Assets/samples.xml
@@ -1,3 +1,1 @@
-<samples>
-  <sample name="Knots" X="10" Y="10" Z="6" periodic="True" ground="empty" screenshots="2"/>
-</samples>
+GenericPropertyJSON:{"name":"voxelTypes","type":-1,"arraySize":7,"arrayType":"VoxelType","children":[{"name":"Array","type":-1,"arraySize":7,"arrayType":"VoxelType","children":[{"name":"size","type":12,"val":7},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"bad6a165ab43a2b4a842cd9cd20ffb1f\",\"localId\":8657189267615039759,\"type\":3,\"instanceID\":27888}"},{"name":"name","type":3,"val":"Empty"},{"name":"symmetry","type":7,"val":"Enum:X"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"77fc479d0a2cb824d908b682b306ca4d\",\"localId\":1330280528381846513,\"type\":3,\"instanceID\":27890}"},{"name":"name","type":3,"val":"Ground"},{"name":"symmetry","type":7,"val":"Enum:X"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"6640120548b5d0641a1cd4bfd10191d9\",\"localId\":8147017978566268856,\"type\":3,\"instanceID\":27892}"},{"name":"name","type":3,"val":"Turn"},{"name":"symmetry","type":7,"val":"Enum:L"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"6a48e8121ea0eb24b9a5401a0682e8f2\",\"localId\":1508886160661466985,\"type\":3,\"instanceID\":27894}"},{"name":"name","type":3,"val":"Down"},{"name":"symmetry","type":7,"val":"Enum:T"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"cbe07edf8ad60cd4ea99fbdd5837a9f7\",\"localId\":5891977651945633426,\"type\":3,\"instanceID\":27896}"},{"name":"name","type":3,"val":"Up"},{"name":"symmetry","type":7,"val":"Enum:T"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"6f8da44b949c80446b5fe7eea83b0a03\",\"localId\":3564089449883909529,\"type\":3,\"instanceID\":27898}"},{"name":"name","type":3,"val":"Vertical"},{"name":"symmetry","type":7,"val":"Enum:X"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":1}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]},{"name":"data","type":-1,"children":[{"name":"x","type":0,"val":0},{"name":"y","type":0,"val":0},{"name":"z","type":0,"val":0},{"name":"voxelObject","type":5,"val":"UnityEditor.ObjectWrapperJSON:{\"guid\":\"2733fb6d90e5cc845b58af5cf97b105a\",\"localId\":6144670608197194975,\"type\":3,\"instanceID\":27900}"},{"name":"name","type":3,"val":"Line"},{"name":"symmetry","type":7,"val":"Enum:I"},{"name":"connections","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"Array","type":-1,"arraySize":6,"arrayType":"int","children":[{"name":"size","type":12,"val":6},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":1},{"name":"data","type":0,"val":0},{"name":"data","type":0,"val":0}]}]},{"name":"rotation","type":17,"val":"Quaternion(0,0,0,0)"}]}]}]}
